# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Construction Coordination Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Job",
        "operationId": "job_api_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/evidence": {
      "get": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_projects__project_id__evidence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "title": "Response Evidence Api Projects  Project Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/bim/import": {
      "post": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Import Ifc",
        "operationId": "import_ifc_api_projects__project_id__bim_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_ifc_api_projects__project_id__bim_import_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/bim/content": {
      "get": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Bim Content",
        "operationId": "bim_content_api_projects__project_id__bim_content_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimization/fixture": {
      "get": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Optimization Fixture",
        "operationId": "optimization_fixture_api_optimization_fixture_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulingProblem-Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects/{project_id}/optimization": {
      "post": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Optimize",
        "operationId": "optimize_api_projects__project_id__optimization_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulingProblem-Input"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/vision": {
      "post": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Vision",
        "operationId": "vision_api_projects__project_id__vision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_vision_api_projects__project_id__vision_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/semantic/index": {
      "post": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Index Document",
        "operationId": "index_document_api_projects__project_id__semantic_index_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingIndexRequest-Input"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/semantic/search": {
      "post": {
        "tags": [
          "capability-jobs"
        ],
        "summary": "Semantic Search",
        "operationId": "semantic_search_api_projects__project_id__semantic_search_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SemanticMatch"
                  },
                  "title": "Response Semantic Search Api Projects  Project Id  Semantic Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "WAITING_APPROVAL",
              "COMPLETED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ],
            "title": "Status",
            "default": "QUEUED"
          },
          "runtime": {
            "type": "string",
            "title": "Runtime",
            "default": "dbos"
          },
          "runtime_execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Execution Id"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation",
            "default": 0
          },
          "runtime_generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Runtime Generation",
            "default": 0
          },
          "category": {
            "type": "string",
            "enum": [
              "coordination",
              "investigation",
              "engineering_recheck",
              "document_parse",
              "bim_import",
              "optimization",
              "vision",
              "embedding_index"
            ],
            "title": "Category",
            "default": "coordination"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "event_id",
          "status",
          "runtime",
          "runtime_execution_id",
          "generation",
          "runtime_generation",
          "category",
          "analysis_id",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "AgentRun"
      },
...
      "Body_import_ifc_api_projects__project_id__bim_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_ifc_api_projects__project_id__bim_import_post"
      },
...
      "Body_vision_api_projects__project_id__vision_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_vision_api_projects__project_id__vision_post"
      },
...
      "CapabilityJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By"
          },
          "request": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DocumentImport"
              },
              {
                "$ref": "#/components/schemas/BIMImport"
              },
              {
                "$ref": "#/components/schemas/OptimizationRequest"
              },
              {
                "$ref": "#/components/schemas/VisionRequest"
              },
              {
                "$ref": "#/components/schemas/EmbeddingIndexRequest-Output"
              }
            ],
            "title": "Request",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "bim_import": "#/components/schemas/BIMImport",
                "document_parse": "#/components/schemas/DocumentImport",
                "embedding_index": "#/components/schemas/EmbeddingIndexRequest-Output",
                "optimization": "#/components/schemas/OptimizationRequest",
                "vision": "#/components/schemas/VisionRequest"
              }
            }
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "requested_by",
          "request",
          "snapshot_id",
          "result"
        ],
        "title": "CapabilityJob"
      },
...
      "EmbeddingIndexRequest-Input": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "embedding_index",
            "title": "Kind",
            "default": "embedding_index"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "chunk_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 128,
            "title": "Chunk Ids",
            "default": []
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "EmbeddingIndexRequest"
      },
...
      "Evidence": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_revision": {
            "type": "string",
            "title": "Source Revision"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "work_package_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Package Id"
          },
          "element_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Element Ids",
            "default": []
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "fact": {
            "type": "string",
            "title": "Fact"
          },
          "quality": {
            "type": "string",
            "enum": [
              "structured",
              "extracted",
              "inferred"
            ],
            "title": "Quality",
            "default": "structured"
          },
          "source_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision Id"
          },
          "viewer_target": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DrawingTarget"
                  },
                  {
                    "$ref": "#/components/schemas/CadTarget"
                  },
                  {
                    "$ref": "#/components/schemas/BimTarget"
                  },
                  {
                    "$ref": "#/components/schemas/DocumentTarget"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "bim": "#/components/schemas/BimTarget",
                    "cad": "#/components/schemas/CadTarget",
                    "document": "#/components/schemas/DocumentTarget",
                    "drawing": "#/components/schemas/DrawingTarget"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "snapshot_id",
          "provider",
          "source_id",
          "source_revision",
          "observed_at",
          "work_package_id",
          "element_ids",
          "page",
          "location",
          "fact",
          "quality",
          "source_revision_id",
          "viewer_target"
        ],
        "title": "Evidence"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizationRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "optimization",
            "title": "Kind",
            "default": "optimization"
          },
          "problem": {
            "$ref": "#/components/schemas/SchedulingProblem-Output"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "problem"
        ],
        "title": "OptimizationRequest"
      },
...
      "SchedulingProblem-Input": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/ScheduleTask-Input"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Tasks"
          },
          "crews": {
            "items": {
              "$ref": "#/components/schemas/Crew-Input"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Crews"
          },
          "equipment_capacities": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Equipment Capacities"
          },
          "horizon": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Horizon",
            "default": 100
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tasks",
          "crews"
        ],
        "title": "SchedulingProblem"
      },
      "SchedulingProblem-Output": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/ScheduleTask-Output"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Tasks"
          },
          "crews": {
            "items": {
              "$ref": "#/components/schemas/Crew-Output"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Crews"
          },
          "equipment_capacities": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Equipment Capacities"
          },
          "horizon": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Horizon",
            "default": 100
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tasks",
          "crews",
          "equipment_capacities",
          "horizon"
        ],
        "title": "SchedulingProblem"
      },
      "SemanticMatch": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "source_hash": {
            "type": "string",
            "title": "Source Hash"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "dimensions": {
            "type": "integer",
            "title": "Dimensions"
          },
          "test_only": {
            "type": "boolean",
            "title": "Test Only"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "chunk_id",
          "text",
          "page",
          "source_hash",
          "score",
          "model",
          "model_version",
          "dimensions",
          "test_only"
        ],
        "title": "SemanticMatch"
      },
      "SemanticQuery": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "source_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Hash"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SemanticQuery"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}